swap to disk: use glibc's bin output

+++ nix-modules/swap-to-disk.nix
@@ -32,18 +32,24 @@
     };
     script = ''
       set -eux
-      ${pkgs.kmod}/bin/modprobe raid0
+      if [ ! -e /dev/md/spill.decrypted ]; then
+        ${pkgs.kmod}/bin/modprobe raid0
 
-      echo 2 > /sys/module/raid0/parameters/default_layout
+        echo 2 > /sys/module/raid0/parameters/default_layout
 
-      ${pkgs.util-linux}/bin/lsblk -d -e 1,7,11,230 -o PATH -n | ${pkgs.findutils}/bin/xargs ${pkgs.mdadm}/bin/mdadm /dev/md/spill.decrypted --create --level=0 --force --raid-devices=$(${pkgs.util-linux}/bin/lsblk -d -e 1,7,11,230 -o PATH -n | ${pkgs.busybox}/bin/wc -l)
+        ${pkgs.util-linux}/bin/lsblk -d -e 1,7,11,230 -o PATH -n > disklist
+        ${pkgs.coreutils}/bin/cat disklist
+        ${pkgs.coreutils}/bin/cat disklist | ${pkgs.findutils}/bin/xargs ${pkgs.mdadm}/bin/mdadm /dev/md/spill.decrypted --create --level=0 --force --raid-devices=$(${pkgs.coreutils}/bin/cat disklist | ${pkgs.busybox}/bin/wc -l)
+        rm disklist
+      fi
+      
       ${pkgs.cryptsetup}/bin/cryptsetup -c aes-xts-plain64 -d /dev/random create spill.encrypted /dev/md/spill.decrypted
 
       ${pkgs.util-linux}/bin/mkswap /dev/mapper/spill.encrypted
       ${pkgs.util-linux}/bin/swapon /dev/mapper/spill.encrypted
 
       size=$(${pkgs.util-linux}/bin/lsblk --noheadings --bytes --output SIZE /dev/mapper/spill.encrypted)
-      pagesize=$(${pkgs.glibc}/bin/getconf PAGESIZE)
+      pagesize=$(${pkgs.glibc.bin}/bin/getconf PAGESIZE)
       inodes=$((size / pagesize))
       ${pkgs.util-linux}/bin/mount -o remount,size=$size,nr_inodes=$inodes /
     '';
